pyproject.toml
[project]
name = "asr-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.9.0",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "pinecone[grpc]>=6.0.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
]
